requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-generativeai==0.8.4
pydantic==2.10.4
python-dotenv==1.0.1
httpx==0.28.1
jinja2==3.1.5
python-multipart==0.0.20
Pillow
mem0ai
qdrant-client
reka-api
yutori
pytest
pytest-asyncio
